+++ BookReader/ErrorMessage.swift
@@ -1,0 +1,13 @@
+//
+//  ErrorMessage.swift
+//  BookReader
+//
+//  Created by Sunrizz on 27.09.2024.
+//
+
+import Foundation
+
+struct ErrorMessage: Identifiable {
+    let id = UUID()
+    let message: String
+}
